+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Todo Tree Change Log
 
+- fix ripgrep link in README.md (thanks to [pke](https://github.com/pke))
+
 ## v0.0.182 - 2020-10-03
 
 - Fix hidden file test